requirements.txt
python-telegram-bot==21.5
openpyxl==3.1.5
pypdf==4.3.1
Pillow==10.4.0
python-dotenv==1.0.1
